[== APRIL TAGS ==] * Added new equation for calculateArmAngleToShoot (has yet to be tested)
- Andrew Maney

+++ src/main/java/frc/robot/Auto.java
@@ -30,9 +30,9 @@ public class Auto {
     private int armStatus = Robot.CONT;
     private int status = Robot.CONT;
     
-    private final float SHOOT1_ANGLE = 339f;   // Up against speaker
+    private final double SHOOT1_ANGLE = 339;   // Up against speaker
     //private final int SHOOT1_ANGLE = 349;   // Edge of safe zone
-    private final float SHOOT2_ANGLE = 339f;
+    private final double SHOOT2_ANGLE = 339;
     
     private double apriltagShootAngle = SHOOT1_ANGLE;   // Start at against the speaker
